requirements.txt
django-graphql-jwt>=0.4.0
Django>=4.2,<5.0
djangorestframework-simplejwt>=5.3
djangorestframework>=3.15
graphene-django>=3.0
gunicorn>=21.2
lxml>=4.9
psycopg2-binary>=2.9
python-dotenv>=1.0
six>=1.16.0
spyne>=2.14.0
